move RouteContext from result package to route package
