+++ sharding-core/sharding-core-rewrite/src/main/java/org/apache/shardingsphere/sharding/rewrite/sql/ShardingSQLBuilder.java
@@ -18,8 +18,8 @@
 package org.apache.shardingsphere.sharding.rewrite.sql;
 
 import com.google.common.base.Optional;
-import org.apache.shardingsphere.underlying.route.result.RouteUnit;
-import org.apache.shardingsphere.underlying.route.result.TableUnit;
+import org.apache.shardingsphere.underlying.route.RouteUnit;
+import org.apache.shardingsphere.underlying.route.TableUnit;
 import org.apache.shardingsphere.core.rule.BindingTableRule;
 import org.apache.shardingsphere.core.rule.ShardingRule;
 import org.apache.shardingsphere.underlying.rewrite.context.SQLRewriteContext;
