+++ sharding-core/sharding-core-route/src/main/java/org/apache/shardingsphere/core/route/type/complex/CartesianRoutingEngine.java
@@ -23,9 +23,9 @@ import com.google.common.collect.Lists;
 import com.google.common.collect.Sets;
 import lombok.RequiredArgsConstructor;
 import org.apache.shardingsphere.core.route.type.RoutingEngine;
-import org.apache.shardingsphere.underlying.route.result.RouteResult;
-import org.apache.shardingsphere.underlying.route.result.TableUnit;
-import org.apache.shardingsphere.underlying.route.result.RouteUnit;
+import org.apache.shardingsphere.underlying.route.RouteResult;
+import org.apache.shardingsphere.underlying.route.TableUnit;
+import org.apache.shardingsphere.underlying.route.RouteUnit;
 
 import java.util.ArrayList;
 import java.util.Collection;
