+++ sharding-core/sharding-core-route/src/test/java/org/apache/shardingsphere/core/route/router/masterslave/MasterSlaveRouteDecoratorTest.java
@@ -19,11 +19,11 @@ package org.apache.shardingsphere.core.route.router.masterslave;
 
 import com.google.common.base.Optional;
 import com.google.common.collect.Lists;
-import org.apache.shardingsphere.underlying.route.result.RouteContext;
+import org.apache.shardingsphere.underlying.route.RouteContext;
 import org.apache.shardingsphere.core.route.ShardingRouteContext;
-import org.apache.shardingsphere.underlying.route.result.RouteResult;
-import org.apache.shardingsphere.underlying.route.result.RouteUnit;
-import org.apache.shardingsphere.underlying.route.result.TableUnit;
+import org.apache.shardingsphere.underlying.route.RouteResult;
+import org.apache.shardingsphere.underlying.route.RouteUnit;
+import org.apache.shardingsphere.underlying.route.TableUnit;
 import org.apache.shardingsphere.core.rule.MasterSlaveRule;
 import org.apache.shardingsphere.core.strategy.masterslave.RandomMasterSlaveLoadBalanceAlgorithm;
 import org.apache.shardingsphere.sql.parser.relation.statement.SQLStatementContext;
