+++ sharding-core/sharding-core-route/src/test/java/org/apache/shardingsphere/core/route/type/ignore/IgnoreRoutingEngineTest.java
@@ -17,7 +17,7 @@
 
 package org.apache.shardingsphere.core.route.type.ignore;
 
-import org.apache.shardingsphere.underlying.route.result.RouteResult;
+import org.apache.shardingsphere.underlying.route.RouteResult;
 import org.junit.Test;
 
 import static org.hamcrest.CoreMatchers.instanceOf;
