+++ sharding-core/sharding-core-route/src/test/java/org/apache/shardingsphere/core/route/type/unicast/UnicastRoutingEngineTest.java
@@ -20,7 +20,7 @@ package org.apache.shardingsphere.core.route.type.unicast;
 import org.apache.shardingsphere.api.config.sharding.ShardingRuleConfiguration;
 import org.apache.shardingsphere.api.config.sharding.TableRuleConfiguration;
 import org.apache.shardingsphere.underlying.common.config.exception.ShardingSphereConfigurationException;
-import org.apache.shardingsphere.underlying.route.result.RouteResult;
+import org.apache.shardingsphere.underlying.route.RouteResult;
 import org.apache.shardingsphere.core.rule.ShardingRule;
 import org.junit.Before;
 import org.junit.Test;
