+++ shardingsphere-underlying/shardingsphere-route/src/main/java/org/apache/shardingsphere/underlying/route/RouteResult.java
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-package org.apache.shardingsphere.underlying.route.result;
+package org.apache.shardingsphere.underlying.route;
 
 import com.google.common.base.Optional;
 import lombok.Getter;
